feat!: move font to gen layout
