+++ packages/next/src/layouts/Root/index.tsx
@@ -5,7 +5,6 @@ import { initI18n, rtlLanguages } from '@payloadcms/translations'
 import { RootProvider } from '@payloadcms/ui'
 import '@payloadcms/ui/scss/app.scss'
 import { buildComponentMap } from '@payloadcms/ui/utilities/buildComponentMap'
-import { Merriweather } from 'next/font/google'
 import { headers as getHeaders, cookies as nextCookies } from 'next/headers.js'
 import { createClientConfig, parseCookies } from 'payload'
 import React from 'react'
@@ -16,14 +15,6 @@ import { getRequestTheme } from '../../utilities/getRequestTheme.js'
 import { DefaultEditView } from '../../views/Edit/Default/index.js'
 import { DefaultListView } from '../../views/List/Default/index.js'
 
-const merriweather = Merriweather({
-  display: 'swap',
-  style: ['normal', 'italic'],
-  subsets: ['latin'],
-  variable: '--font-serif',
-  weight: ['400', '900'],
-})
-
 export const metadata = {
   description: 'Generated by Next.js',
   title: 'Next.js',
@@ -31,9 +22,11 @@ export const metadata = {
 
 export const RootLayout = async ({
   children,
+  className,
   config: configPromise,
 }: {
   children: React.ReactNode
+  className?: string
   config: Promise<SanitizedConfig>
 }) => {
   const config = await configPromise
@@ -100,7 +93,7 @@ export const RootLayout = async ({
   })
 
   return (
-    <html className={merriweather.variable} data-theme={theme} dir={dir} lang={languageCode}>
+    <html className={className} data-theme={theme} dir={dir} lang={languageCode}>
       <body>
         <RootProvider
           componentMap={componentMap}
